refactored helper classes to use case layer
